azure_functions/agents/context/requirements.txt
azure-functions
langchain-openai
pydantic
python-dotenv
supabase
